split UIWidgetsPanelWrapper
